project: Use native assembly optimisations.
The x86 .s files cause vsyasm to crash when compiled in parralel so they needed an output file location set in order to force them to compile one by one.

Any function in an asm file does not export properly so it was required to create a module definition file instead of using dllexport.